requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-dialogflow==2.25.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
idna==3.6
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
urllib3==2.1.0
